requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt


# Flask Framework
Flask==0.10.1

# server
gunicorn==19.7.1
waitress==1.0.2

# Flask Packages
Flask-SQLAlchemy==2.0
psycopg2==2.7.3.1
flask-migrate==2.1.1
Flask-API==1.0
flask-bcrypt==0.7.1
PyJWT==1.5.2
Flask-Script==2.0.5
Flask-Mail=0.9.1

# testing
nose==1.3.7
coveralls==1.2.0

#documentation
flask-cors==3.0.3
flasgger==0.7.0
